cleanup import strcuture
