+++ apps/sim/executor/execution/snapshot-serializer.ts
@@ -1,9 +1,6 @@
 import type { DAG } from '@/executor/dag/builder'
-import {
-  type ExecutionMetadata,
-  ExecutionSnapshot,
-  type SerializableExecutionState,
-} from '@/executor/execution/snapshot'
+import { ExecutionSnapshot } from '@/executor/execution/snapshot'
+import type { ExecutionMetadata, SerializableExecutionState } from '@/executor/execution/types'
 import type { ExecutionContext, SerializedSnapshot } from '@/executor/types'
 
 function mapFromEntries<T>(map?: Map<string, T>): Record<string, T> | undefined {
