+++ apps/sim/executor/execution/types.ts
@@ -1,6 +1,67 @@
-import type { ExecutionMetadata, SerializableExecutionState } from '@/executor/execution/snapshot'
-import type { BlockState, NormalizedBlockOutput } from '@/executor/types'
+import type { Edge } from 'reactflow'
+import type { BlockLog, BlockState, NormalizedBlockOutput } from '@/executor/types'
 import type { SubflowType } from '@/stores/workflows/workflow/types'
+
+export interface ExecutionMetadata {
+  requestId: string
+  executionId: string
+  workflowId: string
+  workspaceId: string
+  userId: string
+  sessionUserId?: string
+  workflowUserId?: string
+  triggerType: string
+  triggerBlockId?: string
+  useDraftState: boolean
+  startTime: string
+  isClientSession?: boolean
+  pendingBlocks?: string[]
+  resumeFromSnapshot?: boolean
+  workflowStateOverride?: {
+    blocks: Record<string, any>
+    edges: Edge[]
+    loops?: Record<string, any>
+    parallels?: Record<string, any>
+    deploymentVersionId?: string
+  }
+}
+
+export interface SerializableExecutionState {
+  blockStates: Record<string, BlockState>
+  executedBlocks: string[]
+  blockLogs: BlockLog[]
+  decisions: {
+    router: Record<string, string>
+    condition: Record<string, string>
+  }
+  completedLoops: string[]
+  loopExecutions?: Record<string, any>
+  parallelExecutions?: Record<string, any>
+  parallelBlockMapping?: Record<string, any>
+  activeExecutionPath: string[]
+  pendingQueue?: string[]
+  remainingEdges?: Edge[]
+  dagIncomingEdges?: Record<string, string[]>
+  completedPauseContexts?: string[]
+}
+
+export interface IterationContext {
+  iterationCurrent: number
+  iterationTotal: number
+  iterationType: SubflowType
+}
+
+export interface ExecutionCallbacks {
+  onStream?: (streamingExec: any) => Promise<void>
+  onBlockStart?: (blockId: string, blockName: string, blockType: string) => Promise<void>
+  onBlockComplete?: (
+    blockId: string,
+    blockName: string,
+    blockType: string,
+    output: any,
+    iterationContext?: IterationContext
+  ) => Promise<void>
+}
 
 export interface ContextExtensions {
   workspaceId?: string
@@ -32,22 +93,14 @@ export interface ContextExtensions {
     blockId: string,
     blockName: string,
     blockType: string,
-    iterationContext?: {
-      iterationCurrent: number
-      iterationTotal: number
-      iterationType: SubflowType
-    }
+    iterationContext?: IterationContext
   ) => Promise<void>
   onBlockComplete?: (
     blockId: string,
     blockName: string,
     blockType: string,
     output: { input?: any; output: NormalizedBlockOutput; executionTime: number },
-    iterationContext?: {
-      iterationCurrent: number
-      iterationTotal: number
-      iterationType: SubflowType
-    }
+    iterationContext?: IterationContext
   ) => Promise<void>
 }
